% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breast.R
\docType{data}
\name{breastwisconsin1}
\alias{breastwisconsin1}
\title{Breast Cancer Winsconsin dataset}
\format{
An object of class \code{data.frame} with 699 rows and 10 columns.
}
\source{
https://archive.ics.uci.edu/ml/datasets/breast+cancer+wisconsin+(original)
}
\usage{
data(breastwisconsin1)
}
\description{
Breast Cancer Winsconsin dataset
}
\keyword{datasets}
