% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmda.R
\docType{data}
\name{Hmda}
\alias{Hmda}
\title{Home Mortgage Disclosure Act dataset}
\format{
An object of class \code{data.frame} with 2380 rows and 13 columns.
}
\source{
Stock, J. H. and Watson, M. W. (2007). Introduction to Econometrics, 2nd ed. Boston: Addison Wesley.
}
\usage{
data(Hmda)
}
\description{
Home Mortgage Disclosure Act dataset
}
\keyword{datasets}
