% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\docType{data}
\name{nba}
\alias{nba}
\title{nba dataset}
\format{
An object of class \code{data.frame} with 1340 rows and 21 columns.
}
\source{
https://data.world/exercises/logistic-regression-exercise-1
}
\usage{
data(nba)
}
\description{
nba dataset
}
\keyword{datasets}
