% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pima.R
\docType{data}
\name{pima}
\alias{pima}
\title{Pima indian diabetes dataset}
\format{
An object of class \code{data.frame} with 768 rows and 9 columns.
}
\source{
https://sci2s.ugr.es/keel/dataset.php?cod=21
}
\usage{
data(pima)
}
\description{
Pima indian diabetes dataset
}
\keyword{datasets}
