% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral.R
\docType{data}
\name{spiral}
\alias{spiral}
\title{spiral sample data}
\format{
An object of class \code{data.frame} with 803 rows and 3 columns.
}
\usage{
data(spiral)
}
\description{
spiral sample data
}
\keyword{datasets}
